+++ docs/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 ### Changed
+- [Issue 144](https://github.com/aza547/wow-recorder/issues/144) - Use the VideoJS player for playback.
+
 ### Fixed
 
 ## [3.6.2] - 2023-03-12